auto-labeler/MATCH/requirements.txt
torch==1.2.0
torchvision==0.4.0
torchgpipe==0.0.5
click==7.0
ruamel.yaml==0.16.5
numpy==1.16.2
scipy==1.2.1
scikit-learn==0.20.3
gensim==3.7.2
tqdm==4.31.1
joblib==0.13.2
logzero==1.5.0
certifi==2021.5.30
charset-normalizer==2.0.1
configparser==5.0.2
cycler==0.10.0
docker-pycreds==0.4.0
filelock==3.0.12
gdown==3.13.0
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
importlib-metadata==4.6.1
kiwisolver==1.3.1
matplotlib==3.3.4
nlpaug==1.1.7
nltk==3.6.2
pandas==1.1.5
pathtools==0.1.2
Pillow==8.3.1
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml.clib==0.2.6
sentry-sdk==1.3.0
shortuuid==1.0.1
six==1.16.0
smart-open==5.1.0
smmap==4.0.0
subprocess32==3.5.4
typing-extensions==3.10.0.0
urllib3==1.26.6
wandb==0.10.33
zipp==3.5.0